requirements.txt
requests==2.32.3
streamlit==1.41.1
beautifulsoup4==4.12.3
pymupdf==1.25.1
pypdf==5.1.0
python-pptx==1.0.2
langchain==0.3.14
openai==1.59.6
tqdm==4.67.1
qdrant-client==1.12.2
python-dotenv==1.0.1